requirement.txt
PySocks
SecretStorage
browser-cookie3
certifi
cffi
charset-normalizer
cryptography
idna
importlib-metadata
jeepney
keyring
lz4
pbkdf2
pyaes
pycparser
pycryptodome
pyperclip
pywin32-ctypes
requests
urllib3
zipp